Refatoração do Pedido e entidades relacionadas para melhorar a estrutura

+++ src/main/java/com/senac/CadastroClienteSenac/DTOs/PedidoItemDTO/CriarPedidoItemDTO.java
@@ -14,5 +14,6 @@ public class CriarPedidoItemDTO {
 
     @NotNull(message = "Quantidade é obrigatória")
     @Positive(message = "Quantidade deve ser positiva")
-    private Double quantidade;
+    private Integer quantidade;
+
 }
